Start project of Questinarium game
The game is created with the goal of educating players through high school level and above questions, in the form of quests. For each completed quest you will receive a certain number of points. As soon as the player gains enough experience, the level of the player will increase, and with it the complexity of the quests. In addition to the complexity, with increasing level the player will open more and more plot dialogues.